message_handler: Unbreak gerrit regexp
Unbreak the regexp for old-format Gerrit links (without the project
path) by making that specific group optional when matching.

Also, properly nest those regexp groups (Base URL -> "/c/" -> project
path) to get rid of vertical lines (which are a hacky way to allow it
to match "").

The hashtag before /c/ is optional as well (due to the new Gerrit UI),
so ignore that as well.

While we're at it, remove the limit of digits in the change number.